Add files via upload
Prerequisite:-
The prerequisite of this project is the basic knowledge of python.

For this project, we are going to use the Virtual Studio Code IDE and the pygame module. Pygame is a library that is used in creating games in Python. It has four important things.

Game Loop
Events
Sprites
Sound
All of these four topics will be discussed in this tutorial.

When you complete this project, you will be able to create more games with more interesting features and concepts. After completing this project, you will be able to create its executable file which you can share with your friends or even use in interviews for showing your python skills.